// src/app/interface/experience.interface.ts
export interface ExperienceInterface {
    id: string
    exp_position: string
    exp_type: string
    exp_orgname: string
    exp_time: string
    exp_startdate: string
    exp_enddate: any
    exp_description: string
    exp_location: string
    exp_status: string
    createdAt: string
    updatedAt: string
    ownerId: string
  }
  